chore: move templates, rename pong_game to pong, rename urls/routes from `game` to `pong`
