+++ backend/pong/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class GameConfig(AppConfig):
+class PongConfig(AppConfig):
 	default_auto_field = 'django.db.models.BigAutoField'
-	name = 'pong_game'
+	name = 'pong'
