+++ backend/pong/scripts/game/cli_update_score.py
@@ -1,18 +1,23 @@
-import argparse
+from argparse import ArgumentParser
+
 import requests
 
 
-def update_score(api_url, csrfToken, username, contol1, contol2, game_id):
-	print('Test')
+def update_score(api_url, csrfToken, score1, score2, game_id):
+	# print(score1)
+	# print(score2)
+	# print(game_id)
+	if score1 > 10 or score2 > 10:
+		print('10 is the max score')
+		return
 
 	headers = {
 		'Content-Type': 'application/json',
 		'X-CSRFToken': csrfToken,
 	}
 	data = {
-		'username': username,
-		'control1': contol1,
-		'control2': contol2,
+		'score1': score1,
+		'score2': score2,
 		'game_id': game_id,
 	}
 
@@ -29,31 +34,27 @@ def update_score(api_url, csrfToken, username, contol1, contol2, game_id):
 
 
 def main():
-	parser = argparse.ArgumentParser(
-		description='CLI tool to create a game using the CreateGame API.'
-	)
+	parser = ArgumentParser(description='CLI tool to create a game using the CreateGame API.')
 	parser.add_argument(
 		'--url',
 		required=True,
-		help='The API endpoint URL (e.g., http://localhost:8000/game/api/get-gameControl/).',
+		help='The API endpoint URL (e.g., http://localhost:8000/pong/api/get-score/).',
 	)
 	parser.add_argument('--token', required=True, help='Your API authentication token.')
-	parser.add_argument('--username', required=True, help='user to change the controll')
-	parser.add_argument('--control1', required=True, help='score of player1')
-	parser.add_argument('--control2', required=True, help='score of player2')
+	parser.add_argument('--score_player1', required=True, help='score of player1')
+	parser.add_argument('--score_player2', required=True, help='score of player2')
 	parser.add_argument('--game_id', required=True, help='game_id')
 
 	args = parser.parse_args()
 
-	update_score(args.url, args.token, args.username, args.control1, args.control2, args.game_id)
+	update_score(args.url, args.token, args.score_player1, args.score_player2, args.game_id)
 
 
 if __name__ == '__main__':
 	main()
 
-# python3 cli_control_keys.py --url http://localhost:8000/game/api/get-gameControl/ \
+# python3 cli_update_score.py --url http://localhost:8000/pong/api/get-score/ \
 #                           --token your_token \
-#                           --username Jsanger \
-#                           --control1 w_s \
-#                           --control2 w_s \
+#                           --score_player1 5 \
+#                           --score_player2 8 \
 #                           --game_id id
