+++ backend/pong/static/pong/GameWebSocket.js
