+++ backend/pong/templates/pong/game_details.html
@@ -1,4 +1,4 @@
-{% extends "game/game_base.html" %}
+{% extends "pong/game_base.html" %}
 {% load static %}
 <!DOCTYPE html>
 <html lang="en">
