+++ backend/pong/urls.py
@@ -1,8 +1,9 @@
 from django.urls import path
-from . import views
-from .api_views import CreateGameView, ScoreBoardView, ControlKeySetting
 
-app_name = 'game'
+from . import views
+from .api_views import ControlKeySetting, CreateGameView, ScoreBoardView
+
+app_name = 'pong'
 urlpatterns = [
 	path('', views.recent_games, name='recent_games'),
 	path('<int:game_id>/', views.game_details, name='game_details'),
