+++ backend/templates/index.html
@@ -43,7 +43,7 @@
 						<a class="nav-item nav-link" href="{% url 'chat' %}">Chat</a>
 						<!-- @follow-up implement about page -->
 						<a class="nav-item nav-link" href="/quiz/" id="quiz-link">Quiz App</a>
-						<a class="nav-item nav-link" href="{% url 'game:recent_games' %}">Game</a>
+						<a class="nav-item nav-link" href="{% url 'pong:recent_games' %}">Pong</a>
 						<a class="nav-item nav-link" href="/dashboard/" id="dashboard-link">Dashboard</a>
 					</div>
 					<!-- Navbar Right Side -->
