+++ backend/templates/navbar.html
@@ -23,7 +23,7 @@
 					<a class="nav-item nav-link" href="/chat">Chat</a>
 					<a class="nav-item nav-link" href="/about">About</a>
 					<a class="nav-item nav-link" href="/quiz">Quiz App</a>
-					<a class="nav-item nav-link" href="{% url 'game:recent_games' %}">Game</a>
+					<a class="nav-item nav-link" href="{% url 'pong:recent_games' %}">Pong</a>
 				</div>
 				<!-- Navbar Right Side -->
 				<div class="navbar-nav">
