+++ backend/transcendence/asgi.py
@@ -13,7 +13,7 @@ from channels.auth import AuthMiddlewareStack
 from channels.routing import ProtocolTypeRouter, URLRouter
 from chat import routing as chat_routing
 from django.core.asgi import get_asgi_application
-from pong_game import routing as pong_routing
+from pong import routing as pong_routing
 from quiz import routing as quiz_routing
 from user_management import routing as user_status_routing
