+++ backend/transcendence/settings.py
@@ -36,7 +36,7 @@ INTERNAL_IPS = [
 
 INSTALLED_APPS = [
 	'quiz.apps.QuizConfig',
-	'pong_game.apps.GameConfig',
+	'pong.apps.PongConfig',
 	'chat.apps.ChatConfig',
 	'dashboard.apps.DashboardConfig',
 	'channels',
