+++ backend/user_management/templates/users/list.html
@@ -1,4 +1,4 @@
-{% extends "game/game_base.html" %}
+{% extends "pong/game_base.html" %}
 {% load static %}
 {% block content %}
 <!DOCTYPE html>
