+++ backend/user_management/templates/users/public_profile.html
@@ -1,4 +1,4 @@
-{% extends "game/game_base.html" %}
+{% extends "pong/game_base.html" %}
 {% load static %}
 {% block content %}
 <!DOCTYPE html>
@@ -127,7 +127,7 @@
 				<h3>Wins</h3>
 				{% for game in games_won %}
 				<div class="game-item">
-					<a href="{% url 'game:game_details' game.id %}"
+					<a href="{% url 'pong:game_details' game.id %}"
 						class="list-group-item list-group-item-action list-group-item-success" aria-current="true">
 						{{ game.player1 }} vs {{ game.player2 }}
 					</a>
@@ -141,7 +141,7 @@
 				<h3>Losses</h3>
 				{% for game in games_lost %}
 				<div class="game-item">
-					<a href="{% url 'game:game_details' game.id %}"
+					<a href="{% url 'pong:game_details' game.id %}"
 						class="list-group-item list-group-item-action list-group-item-danger" aria-current="true">
 						{{ game.player1 }} vs {{ game.player2 }}
 					</a>
