+++ backend/user_management/views.py
@@ -4,8 +4,8 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth.forms import PasswordChangeForm
 from django.db.models import F
 from django.shortcuts import redirect, render
-from pong_game.models import PongGame
-from pong_game.utils import win_to_loss_ratio
+from pong.models import PongGame
+from pong.utils import win_to_loss_ratio
 
 from user_management.friends import Friends_Manager
 
@@ -68,16 +68,14 @@ def account(request):
 @login_required
 def public_profile(request, query_user):
 	query_user_instance = User.objects.get(username=query_user)
-	pong_games_finished = PongGame.objects.filter(
+	pongs_finished = PongGame.objects.filter(
 		player1=query_user_instance, pending=False
 	) | PongGame.objects.filter(player2=query_user_instance, pending=False)
-	games_won = pong_games_finished.filter(
+	games_won = pongs_finished.filter(
 		player1=query_user_instance, score1__gt=F('score2'), pending=False
-	) | pong_games_finished.filter(
-		player2=query_user_instance, score2__gt=F('score1'), pending=False
-	)
+	) | pongs_finished.filter(player2=query_user_instance, score2__gt=F('score1'), pending=False)
 	# something to use the display_name in games (playing as display_name) @follow-up
-	games_lost = [game for game in pong_games_finished if game not in games_won]
+	games_lost = [game for game in pongs_finished if game not in games_won]
 	games_won = sorted(games_won, key=lambda game: game.played_at, reverse=True)
 	games_lost = sorted(games_lost, key=lambda game: game.played_at, reverse=True)
 	friends = Friends_Manager.fetch_friends_public(user_instance=query_user_instance)
